chore: moved tests in a specific folder
