+++ tests/providers/aws/generateApiGatewayConfiguration.test.js
@@ -1,6 +1,6 @@
 // @ts-nocheck
-const { generateTerraformConfiguration } = require("../src/providers/aws/generateApiGatewayConfiguration");
-const { setConfiguration } = require("../src/configuration");
+const { generateTerraformConfiguration } = require("../../../src/providers/aws/generateApiGatewayConfiguration");
+const { setConfiguration } = require("../../../src/configuration");
 
 describe("generateTerraformConfiguration", () => {
 	it("should generate the correct configuration", () => {
