+++ tests/providers/aws/terraFormGatewayMethod.test.js
@@ -1,6 +1,6 @@
 // @ts-nocheck
-const { setConfiguration } = require("../src/configuration");
-const { generateGatewayMethod } = require("../src/providers/aws/resources/terraFormGatewayMethod");
+const { setConfiguration } = require("../../../src/configuration");
+const { generateGatewayMethod } = require("../../../src/providers/aws/resources/terraFormGatewayMethod");
 describe("terraFormGatewayMethod", () => {
 	it("should create a correct resource", () => {
 		setConfiguration({
